requirements.txt
requests
pyrogram
tgcrypto
pyromod
python-dotenv
duckduckgo-search
requests
flask